% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{as_ISO8601_character}
\alias{as_ISO8601_character}
\title{Coerce input date parameters to ISO 8601 format for Bamboo HR REST API}
\usage{
as_ISO8601_character(time)
}
\description{
Coerce input date parameters to ISO 8601 format for Bamboo HR REST API
}
\examples{
as_ISO8601_character("2020-09-01")
as_ISO8601_character("2020-09-01 00:00:00.001")

}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
