% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambooHR-package.R
\docType{package}
\name{bambooHR-package}
\alias{bambooHR}
\alias{bambooHR-package}
\title{bambooHR: A Wrapper to the 'BambooHR' API}
\description{
Enables a user to consume the 'BambooHR' API endpoints using R. The actual URL of the API will depend on your company domain, and will be handled by the package automatically once you setup the config file. The API documentation can be found here <https://documentation.bamboohr.com/docs>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mangothecat.github.io/bambooHR/}
  \item Report bugs at \url{https://github.com/MangoTheCat/bambooHR/issues}
}

}
\author{
\strong{Maintainer}: Tom Bowling \email{tom.bowling@ascent.io}

Authors:
\itemize{
  \item Harry Alexander \email{harry.alexander@ascent.io}
  \item Tim Fry \email{tim.fry@ascent.io}
}

Other contributors:
\itemize{
  \item Andrew Little [contributor]
  \item Mark Druffel \email{mdruffel@propellerpdx.com} [contributor]
}

}
\keyword{internal}
