% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_url.r
\name{build_url}
\alias{build_url}
\title{fill in}
\usage{
build_url(company_domain, api_version, base_url)
}
\arguments{
\item{company_domain}{tbd}

\item{table_name}{tbd}

\item{employee_id}{tbd}
}
\value{
text
}
\description{
fill in
}
\examples{


}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
