% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{date_formats}
\alias{date_formats}
\alias{is_ymd}
\alias{is_ymd_hms}
\alias{is_ymd_hmsms}
\title{Check Date Format}
\usage{
is_ymd(date = NULL)

is_ymd_hms(date = NULL)

is_ymd_hmsms(date = NULL)
}
\description{
Functions to check date and datetime formats
}
\section{Functions}{
\itemize{
\item \code{is_ymd}: Check format is YYYY-MM-DD

\item \code{is_ymd_hms}: Check format is YYYY-MM-DD HH:MM:SS

\item \code{is_ymd_hmsms}: Check format is YYYY-MM-DD HH:MM:SS.mmm
}}

\keyword{internal}
