% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_file.R
\name{get_company_file}
\alias{get_company_file}
\title{Retrieve A Company File.}
\usage{
get_company_file(file_id = "view", api_version = "v1", suppress_view = FALSE)
}
\arguments{
\item{file_id}{The ID of the file to get from BambooHR.}

\item{api_version}{version of the BambooHR API.}

\item{suppress_view}{prevent display of results when file_id = "view", default is FALSE.}
}
\value{
returns a response object.
}
\description{
`get_company_file` takes a file_id (string) and
api version (string) as arguments and then requests
and returns data about the corresponding file from the BambooHR API,
the file is then downloaded if possible.
}
\examples{
\dontrun{
response <- get_company_file(
"480",
api_version = "v1"
)
}

}
\author{
Harry Alexander, \email{harry.alexander@ascent.io}
}
