% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_employee_file_meta.R
\name{get_employee_file_meta}
\alias{get_employee_file_meta}
\title{Get Employee File Meta Data}
\usage{
get_employee_file_meta(...)
}
\arguments{
\item{...}{tbd}
}
\value{
text
}
\description{
fill in
}
