% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Get Meta Information}
\usage{
get_meta(
  query = c("fields", "tables", "lists", "users", "time_off/types", "time_off/policies"),
  api_version = "v1"
)
}
\arguments{
\item{query}{One of: "fields", "tables", "lists", "users", "time_off/types", "time_off/policies".}

\item{api_version}{(optional) - Version of API to use to make request. Default is "v1".}
}
\value{
A [tibble::tibble()] object.
}
\description{
Submits a get request to retrieve account meta information that
 can be used to help discover the types of field/table etc. available in your account.
}
\examples{
\dontrun{
res <- get_meta("fields")

res2 <- get_meta("users")
}

}
\references{
\url{https://documentation.bamboohr.com/reference/account-information-1}
}
