% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_request}
\alias{get_request}
\title{Bamboo HR API get request wrapper}
\usage{
get_request(
 
    url = "https://api.bamboohr.com/api/gateway.php/propellerpdx/v1/employees/all/tables/compensation",
  key = keys$BAMBOO_KEY,
  verbose
)
}
\arguments{
\item{url}{Character url of the get request.}

\item{key}{Character key. Register in your account, see \href{https://documentation.bamboohr.com/docs#section-authentication}{authentication} for details.}

\item{verbose}{Logical scalar. Should the function provide verbose messaging back on each step?}
}
\value{
response object
}
\description{
Submits get requests to the Bamboo API
}
\examples{


}
\references{
\url{https://github.com/r-lib/httr}
\url{https://curl.haxx.se/libcurl/c/CURLOPT_HTTP_VERSION.html}
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
