% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeoff.R
\name{get_timeoff}
\alias{get_timeoff}
\title{Get Time Off Requests}
\usage{
get_timeoff(
  start,
  end,
  id = NULL,
  action = "view",
  employee_id = NULL,
  type = NULL,
  status = c("approved", "denied", "superseded", "requested", "canceled"),
  url_args = NULL
)
}
\arguments{
\item{start}{Character in format "YYYY-MM-DD". Only show time off that occurs on/after the specified start date.}

\item{end}{Character in format "YYYY-MM-DD". Only show time off that occurs on/before the specified end date.}

\item{id}{(optional) Integer - A particular request ID to limit the response to.}

\item{action}{(optional) - Limit to requests that the user has a particular level of access to. Legal values are: "view" or "approve".}

\item{employee_id}{(optional) Character - A particular employee ID to limit the response to.}

\item{type}{(optional) - A vector of time off types IDs to include limit the response to. Default is all types are included.}

\item{status}{(optional) - A vector of request status values to include. Legal values are "approved", "denied", "superseded", "requested", "canceled". Default is all status types.}

\item{url_args}{(optional) - A list of named or unnamed arguments to be passed to build_URL function.}
}
\value{
A \code{\link[httr:response]{httr::response()}} object.
}
\description{
Get Time Off Requests
}
\examples{
\dontrun{
# find valid types

type_list <- get_timeoff_types()
out <- httr::content(type_list, "parsed")
types <- sapply(out$timeOffTypes, "[[", "id")

res <- get_timeoff("2022-01-01", "2022-02-01", type = types)
httr::content(res, "parsed")


res <- get_timeoff("2022-01-01", "2022-02-01", action = "approve", status = c("approved", "denied"))
httr::content(res, "parsed")

res <- get_timeoff("2022-01-01", "2022-02-01", url_args = list(company_domain = "company"))
httr::content(res, "parsed")
}
}
\references{
\url{https://documentation.bamboohr.com/reference/time-off-1}
}
