% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeoff.R
\name{get_timeoff_policies}
\alias{get_timeoff_policies}
\title{Get Time Off policies}
\usage{
get_timeoff_policies()
}
\value{
A \code{\link[httr:response]{httr::response()}} object where the content is a JSON format that contains the
list of time off policies.
}
\description{
Wrapper to get_meta function that returns a list of time off policies.
}
\note{
Currently do not have the permissions to test this so returning the raw response object
}
\examples{
\dontrun{
res <- get_timeoff_policies()
httr::content(res, "parsed")
}
}
\references{
\url{https://documentation.bamboohr.com/reference/get-time-off-policies}
}
