% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeoff.R
\name{get_timeoff_types}
\alias{get_timeoff_types}
\title{Get Time Off Types}
\usage{
get_timeoff_types()
}
\value{
A vector of type ids
}
\description{
Wrapper to get_meta function that returns time off types
}
\examples{
\dontrun{
types <- get_timeoff_types()
type_ids <- types \%>\% tidyr::drop_na(id) \%>\%  dplyr::pull(id)
}
}
\references{
\url{https://documentation.bamboohr.com/reference/get-time-off-types}
}
