% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeoff.R
\name{get_whos_out}
\alias{get_whos_out}
\title{Get a list of Who's Out}
\usage{
get_whos_out(start = "", end = "", api_version = "v1")
}
\arguments{
\item{start}{(optional) - a date in the form YYYY-MM-DD - defaults to the current date.}

\item{end}{(optional) - a date in the form YYYY-MM-DD - defaults to 14 days from the start date.}

\item{api_version}{(optional) - Version of API to use to make request. Default is "v1".}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} object.
}
\description{
Get a list of Who's Out
}
\examples{
\dontrun{
res <- get_whos_out()

res2 <- get_whos_out(start = "2022-01-12")

res3 <- get_whos_out(start = "2022-01-01", end = "2022-04-01")
}
}
\references{
\url{https://documentation.bamboohr.com/reference/get-a-list-of-whos-out-1}
}
