% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_view.R
\name{response_view}
\alias{response_view}
\title{View Data About a Response.}
\usage{
response_view(x, title)
}
\arguments{
\item{x}{An R object which can be coerced to a data frame with non-zero numbers of rows and columns.}

\item{title}{Title for the viewer window.}
}
\value{
NULL.
}
\description{
This function uses the `View` function from Rstudio if available,
otherwise it uses the utils::View function.
}
