% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambooHR-package.R
\docType{package}
\name{bambooHR-package}
\alias{bambooHR}
\alias{bambooHR-package}
\title{bambooHR: An R wrapper to the BambooHR API}
\description{
Enables a user to consume the BambooHR API endpoints using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mangothecat.github.io/bambooHR}
}

}
\author{
\strong{Maintainer}: Tom Bowling \email{tom.bowling@ascent.io}

Authors:
\itemize{
  \item Harry Alexander \email{harry.alexander@ascent.io}
  \item Tim Fry \email{tim.fry@ascent.io}
  \item Mark Druffel \email{mdruffel@propellerpdx.com}
}

}
\keyword{internal}
