% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_file.R
\name{get_company_file}
\alias{get_company_file}
\title{Retrieve A Company File}
\usage{
get_company_file(file_id, ...)
}
\arguments{
\item{file_id}{tbd}

\item{...}{tbd}
}
\value{
text
}
\description{
fill in
}
