% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_employee.r
\name{get_employee}
\alias{get_employee}
\title{Get employee data by specifying a set of fields.}
\usage{
get_employee(
  id = "directory",
  fields = NULL,
  config = "~/.bambooHR_user_config.json"
)
}
\arguments{
\item{id}{Employee ID (character or numeric). The default of "directory" will give a directory of all
employees, along with a table of possible fields. The special employee ID of
zero (0) means to use the employee ID associated with the API key (if any).}

\item{fields}{Character vector of fields to return. To see available fields, run
\code{get_employee()}.}

\item{config}{File path to config file. This may be useful if you want to swap
between multiple company domains and API keys. Note that this will}
}
\value{
response object
}
\description{
This is suitable for getting basic employee information, including current
values for fields that are part of a historical table, like job title,
or compensation information. See the fields endpoint for a list of possible
 fields.
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
