% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_employee_file.R
\name{get_employee_file}
\alias{get_employee_file}
\title{Retrieve An Employee File}
\usage{
get_employee_file(
  id = "0",
  file_id = "view",
  api_version = "v1",
  suppress_view = FALSE
)
}
\arguments{
\item{id}{The employee id of the employee.}

\item{file_id}{The file id of the file to be returned.}

\item{api_version}{The version of BambooHR API to be used.}

\item{suppress_view}{prevent display of results when file_id = "view", default is FALSE.}
}
\value{
returns a response object.
}
\description{
`get_employee_file` takes 'id' (string), 'file_id', (string), and
'api_version' (string) and then requests and returns data about
the corresponding file from the BambooHR API.
The file will then be written to the user's working directory if possible.
}
\examples{
\dontrun{
response <- get_employee_file(
id = 0,
file_id = "480",
api_version = "v1"
)
}

}
\author{
Harry Alexander, \email{harry.alexander@ascent.io}
}
