% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Bamboo API get request wrapper for BambooHR tables}
\usage{
get_meta(data)
}
\arguments{
\item{data}{fields, tables, lists, users, time_off/types, time_off/policies}

\item{...}{tbd}
}
\value{
tbl_df
}
\description{
Submits a get request to retrieve a specified table from the system, valid tables are retreived from the
meta data API
}
\references{
\url{https://documentation.bamboohr.com/reference#get-employee-table-row-1}
}
