% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report.R
\name{get_report}
\alias{get_report}
\title{Get Company Report}
\usage{
get_report(
  id,
  format = c("CSV", "PDF", "XLS", "XML", "JSON"),
  field_filtering = TRUE,
  only_current = TRUE,
  api_version = "v1"
)
}
\arguments{
\item{id}{The report ID. You can get the report number by hovering over the
report name on the reports page and grabbing the ID.}

\item{format}{The output format for the report. Supported formats: CSV, PDF,
XLS, XML, JSON.}

\item{field_filtering}{Logical, default [TRUE]. Whether to apply duplicate
field filtering.}

\item{only_current}{Logical default [TRUE]. Should the report be limited
to only current employees.}

\item{api_version}{(optional) - Version of API to use to make request. Default is "v1".}
}
\value{
Currently the function simply returns a parsed response object.
}
\description{
Request one of your pre-existing company reports from the reporting section.
At present, only reports from the Company Reports section are supported.
}
