% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report.R
\name{get_report}
\alias{get_report}
\title{Bamboo HR API get request wrapper}
\usage{
get_report(id, format, fields, ...)
}
\arguments{
\item{id}{employee ID. The special employee ID of zero (0) means to use the employee ID associated with the API key (if any).}

\item{format}{the format for query values appended to URL.}

\item{fields}{vector of values}

\item{verbose}{Logical scalar. Should the function provide verbose messaging back on each step?}
}
\value{
response object
}
\description{
Submits get requests to the Bamboo API
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
