% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeoff.R
\name{get_timeoff_types}
\alias{get_timeoff_types}
\title{Get Time Off Types}
\usage{
get_timeoff_types()
}
\value{
A \code{\link[httr:response]{httr::response()}} object where the content is a JSON format that contains the
list of time off types the user has permissions on.
}
\description{
Wrapper to get_meta function that returns time off types
}
\examples{
\dontrun{
res <- get_timeoff_types()
httr::content(res, "parsed")
}
}
\references{
\url{https://documentation.bamboohr.com/reference/get-time-off-types}
}
