% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeoff.R
\name{get_whos_out}
\alias{get_whos_out}
\title{Get a list of Who's Out}
\usage{
get_whos_out(start = "", end = "", url_args = NULL)
}
\arguments{
\item{start}{(optional) - a date in the form YYYY-MM-DD - defaults to the current date.}

\item{end}{(optional) - a date in the form YYYY-MM-DD - defaults to 14 days from the start date.}

\item{url_args}{(optional) - A list of named or unnamed arguments to be passed to build_URL function.}
}
\value{
A \code{\link[httr:response]{httr::response()}} object where the content is a JSON format that contains a list,
sorted by date, of employees who will be out, and company holidays, for a period of time.
}
\description{
Get a list of Who's Out
}
\examples{
\dontrun{
res <- get_whos_out()
httr::content(res, "parsed")

res <- get_whos_out(start = "2022-01-12")
httr::content(res, "parsed")

res <- get_whos_out(start = "2022-01-01", end = "2022-04-01")
httr::content(res, "parsed")
}
}
\references{
\url{https://documentation.bamboohr.com/reference/get-a-list-of-whos-out-1}
}
